# video-pipeline/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AI Video Series Workflow Control Plane API
  version: 1.0.0
  description: |
    Contract-first API for immutable video-production revisions, approvals,
    Temporal operations, generation runs, cancellation/recovery, stale impact,
    and lineage manifests. FLO-103 implements the Series/Source,
    GenerationPlan, Episode production batch, GenerationRun
    pause/cancel/resume/query, Approval, Operation, stale-impact, and Manifest
    endpoints against PostgreSQL and Temporal. Provider administration and
    direct provider-job/callback operations are intentionally not public
    control-plane endpoints: the Temporal worker owns that internal adapter
    boundary and persists its job/callback reconciliation projections.
servers:
  - url: http://localhost:18080
    description: Local video control plane
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/generation-plans:
    post:
      tags: [Production]
      operationId: createGenerationPlan
      description: |
        Builds a Dry-run task graph and estimate without submitting a paid
        provider request. Confirmation of this immutable plan is required by
        provider-job submission.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGenerationPlanRequest"
      responses:
        "201":
          description: Immutable plan, estimate, and budget decision inputs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationPlan"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/PolicyBlocked"
  /api/v1/series:
    post:
      tags: [Series]
      operationId: createSeries
      x-fr: [FR-P0-01]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSeriesRequest"
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/PolicyBlocked"
  /api/v1/series/{seriesId}/sources:
    post:
      tags: [Series]
      operationId: createSourceRevision
      x-fr: [FR-P0-02]
      parameters:
        - $ref: "#/components/parameters/SeriesID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/ExpectedRevision"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceRevisionRequest"
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/PolicyBlocked"
  /api/v1/sources/{sourceRevisionId}/compilations:
    post:
      tags: [Series]
      operationId: startContentCompilation
      x-fr: [FR-P0-02, FR-P0-03, FR-P0-04]
      description: |
        Starts provider-neutral structured extraction and dramatization from one
        authorized immutable source revision. Every returned world, character,
        relationship, location, prop, episode, scene, and shot revision remains
        editable only by appending a child revision.
      parameters:
        - $ref: "#/components/parameters/SourceRevisionID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartContentCompilationRequest"
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/PolicyBlocked"
  /api/v1/series/{seriesId}/impacts:
    get:
      tags: [Lineage]
      operationId: listRevisionImpacts
      x-fr: [FR-P0-21]
      parameters:
        - $ref: "#/components/parameters/SeriesID"
        - name: sourceRevisionId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Direct and transitive consumers with freshness state
          content:
            application/json:
              schema:
                type: object
                required: [sourceRevisionId, impacts]
                properties:
                  sourceRevisionId:
                    type: string
                    format: uuid
                  impacts:
                    type: array
                    items:
                      $ref: "#/components/schemas/FreshnessImpact"
  /api/v1/episodes/{episodeId}/production-batches:
    post:
      tags: [Production]
      operationId: startEpisodeProduction
      x-fr: [FR-P0-08, FR-P0-09, FR-P0-10, FR-P0-11, FR-P0-12]
      parameters:
        - $ref: "#/components/parameters/EpisodeID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/ExpectedRevision"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartProductionRequest"
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/PolicyBlocked"
  /api/v1/shots/{shotId}/runs:
    post:
      tags: [Production]
      operationId: createGenerationRun
      x-fr: [FR-P0-11, FR-P0-14]
      parameters:
        - $ref: "#/components/parameters/ShotID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/ExpectedRevision"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGenerationRunRequest"
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        "303":
          description: An immutable successful artifact already exists for runSpecDigest
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/PolicyBlocked"
...
  /api/v1/runs/{runId}:
    get:
      tags: [Production]
      operationId: getGenerationRun
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Current queryable run projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationRun"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/runs/{runId}/pause:
    post:
      tags: [Production]
      operationId: pauseGenerationRun
      x-fr: [FR-P0-15]
      parameters:
        - $ref: "#/components/parameters/RunID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/ExpectedRevision"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [actor, reasonCode]
              properties:
                actor:
                  $ref: "#/components/schemas/Actor"
                reasonCode:
                  type: string
                  minLength: 1
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/runs/{runId}/cancel:
    post:
      tags: [Production]
      operationId: cancelGenerationRun
      x-fr: [FR-P0-16]
      parameters:
        - $ref: "#/components/parameters/RunID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/ExpectedRevision"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [actor, reasonCode]
              properties:
                actor:
                  $ref: "#/components/schemas/Actor"
                reasonCode:
                  type: string
                  minLength: 1
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/runs/{runId}/resume:
    post:
      tags: [Production]
      operationId: resumeGenerationRun
      x-fr: [FR-P0-15]
      parameters:
        - $ref: "#/components/parameters/RunID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/ExpectedRevision"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [actor, recoveryMode]
              properties:
                actor:
                  $ref: "#/components/schemas/Actor"
                recoveryMode:
                  type: string
                  enum: [RESUME_PAUSED, RECONCILE_HISTORY, RETRY_INFRASTRUCTURE]
      responses:
        "202":
          $ref: "#/components/responses/OperationAccepted"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/PolicyBlocked"
  /api/v1/runs/{runId}/publication-lock:
    post:
      tags: [Approvals, Lineage]
      operationId: lockRunForPublication
      x-fr: [FR-P0-20, FR-P0-22, FR-P0-23]
      description: |
        Creates one immutable publication lock only when the run succeeded, the
        QC report passed and matches, and G3 approved that exact manifest.
      parameters:
        - $ref: "#/components/parameters/RunID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LockPublicationRequest"
      responses:
        "201":
          description: QC- and G3-bound immutable publication lock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationLock"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/PolicyBlocked"
  /api/v1/approvals:
    post:
      tags: [Approvals]
      operationId: createApprovalDecision
      x-fr: [FR-P0-05, FR-P0-08, FR-P0-20, FR-P0-23]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApprovalDecisionRequest"
      responses:
        "201":
          description: Immutable approval decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDecision"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/PolicyBlocked"
  /api/v1/manifests/{scopeType}/{revisionId}:
    get:
      tags: [Lineage]
      operationId: getGenerationManifest
      x-fr: [FR-P0-22]
      parameters:
        - name: scopeType
          in: path
          required: true
          schema:
            type: string
            enum: [SHOT, EPISODE]
        - name: revisionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Immutable manifest and verification hash
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationManifest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/operations/{operationId}:
    get:
      tags: [System]
      operationId: getOperation
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Long-running operation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Production accepts HS256 tokens with sub, role, aud=video-control-plane,
        and an unexpired exp. Mutation request actor must match signed sub/role.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated UUID. Reuse with the same canonical request returns the
        original result; reuse with a different request hash returns 409.
      schema:
        type: string
        format: uuid
    ExpectedRevision:
      name: If-Match
      in: header
      required: true
      description: Strong aggregate revision, encoded as a quoted integer ETag.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
    SeriesID:
      name: seriesId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SourceRevisionID:
      name: sourceRevisionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EpisodeID:
      name: episodeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ShotID:
      name: shotId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RunID:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    OperationAccepted:
      description: Durable operation accepted
      headers:
        Location:
          schema:
            type: string
            format: uri-reference
        ETag:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Operation"
    ValidationError:
      description: Request does not satisfy the stable schema
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Revision, idempotency, active-run, or terminal-state conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    PolicyBlocked:
      description: Gate, freshness, profile, license, consent, budget, or capability blocked the operation
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Object not found or not visible to the actor
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    CreateGenerationPlanRequest:
      type: object
      additionalProperties: false
      required:
        [schemaVersion, seriesId, episodeRevisionId, shotSpecRevisionIds, candidatesPerShot, routeSnapshot, budgetLimit, executionPolicy, actor]
      properties:
        schemaVersion:
          const: v1
        seriesId:
          type: string
          format: uuid
        episodeRevisionId:
          type: string
          format: uuid
          description: Exact immutable episode revision covered by the content safety decision.
        shotSpecRevisionIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            format: uuid
        candidatesPerShot:
          type: integer
          minimum: 1
        routeSnapshot:
          $ref: "#/components/schemas/ModelRouteSnapshot"
        budgetLimit:
          $ref: "#/components/schemas/BudgetLimit"
        speechBudgetLimit:
          allOf:
            - $ref: "#/components/schemas/BudgetLimit"
          description: Optional immutable TTS envelope; required when post-production speech is enabled.
        executionPolicy:
          $ref: "#/components/schemas/ExecutionPolicy"
        actor:
          $ref: "#/components/schemas/Actor"
    GenerationPlan:
      type: object
      additionalProperties: false
      required:
        [generationPlanId, state, dryRun, shotCount, providerCallCount, routeSnapshot, executionPolicy, estimate, budgetDecision, planHash]
      properties:
        generationPlanId:
          type: string
          format: uuid
        state:
          enum: [DRAFT, READY_FOR_CONFIRMATION, BLOCKED]
        dryRun:
          type: boolean
          const: true
        shotCount:
          type: integer
          minimum: 1
        providerCallCount:
          type: integer
          minimum: 1
        routeSnapshot:
          $ref: "#/components/schemas/ModelRouteSnapshot"
        executionPolicy:
          $ref: "#/components/schemas/ExecutionPolicy"
        estimate:
          $ref: "#/components/schemas/CostEstimate"
        speechBudgetLimit:
          allOf:
            - $ref: "#/components/schemas/BudgetLimit"
          description: Immutable TTS envelope included in the plan hash.
        budgetDecision:
          enum: [WITHIN_BUDGET, AMOUNT_UNKNOWN_REQUIRES_CONFIRMATION, EXCEEDS_BUDGET]
        planHash:
          $ref: "#/components/schemas/SHA256"
...
    Actor:
      type: object
      additionalProperties: false
      required: [actorId, role]
      properties:
        actorId:
          type: string
          minLength: 1
        role:
          type: string
          enum: [CREATOR, DIRECTOR, ARTIST, PRODUCER, REVIEWER, SAFETY_REVIEWER, OPERATOR, ADMIN]
    CreateSeriesRequest:
      type: object
      additionalProperties: false
      required: [schemaVersion, title, generationProfileRevisionId, rightsDeclaration, actor]
      properties:
        schemaVersion:
          const: v1
        title:
          type: string
          minLength: 1
          maxLength: 200
        generationProfileRevisionId:
          type: string
          format: uuid
        rightsDeclaration:
          type: object
          required: [basis, evidenceArtifactHash]
          properties:
            basis:
              type: string
            evidenceArtifactHash:
              $ref: "#/components/schemas/SHA256"
        actor:
          $ref: "#/components/schemas/Actor"
    CreateSourceRevisionRequest:
      type: object
      additionalProperties: false
      required: [schemaVersion, artifactHash, artifactUri, language, rightsSnapshotId, actor]
      properties:
        schemaVersion:
          const: v1
        artifactHash:
          $ref: "#/components/schemas/SHA256"
        artifactUri:
          $ref: "#/components/schemas/ArtifactURI"
        language:
          type: string
        rightsSnapshotId:
          type: string
          format: uuid
        parentRevisionId:
          type: string
          format: uuid
        actor:
          $ref: "#/components/schemas/Actor"
    StartContentCompilationRequest:
      type: object
      additionalProperties: false
      required: [schemaVersion, sourceHash, stages, textRouteSnapshot, actor]
      properties:
        schemaVersion:
          const: v1
        sourceHash:
          $ref: "#/components/schemas/SHA256"
        stages:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum: [STRUCTURE, EPISODES, SCENES, SHOTS]
        textRouteSnapshot:
          $ref: "#/components/schemas/ModelRouteSnapshot"
        actor:
          $ref: "#/components/schemas/Actor"
...
    StartProductionRequest:
      type: object
      additionalProperties: false
      required:
        [schemaVersion, episodeRevisionId, shotSpecRevisionIds, generationProfileRevisionId, gate2DecisionId, generationPlanId, routeSnapshot, budgetApprovalId, executionPolicy, actor]
      properties:
        schemaVersion:
          const: v1
        episodeRevisionId:
          type: string
          format: uuid
        shotSpecRevisionIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            format: uuid
        generationProfileRevisionId:
          type: string
          format: uuid
        gate2DecisionId:
          type: string
          format: uuid
        generationPlanId:
          type: string
          format: uuid
        routeSnapshot:
          $ref: "#/components/schemas/ModelRouteSnapshot"
        budgetApprovalId:
          type: string
          format: uuid
        executionPolicy:
          $ref: "#/components/schemas/ExecutionPolicy"
        postProduction:
          $ref: "#/components/schemas/PostProductionRequest"
        actor:
          $ref: "#/components/schemas/Actor"
...
    CreateGenerationRunRequest:
      type: object
      additionalProperties: false
      required:
        [schemaVersion, shotSpecRevisionId, promptSnapshotId, generationProfileRevisionId, generationPlanId, routeSnapshot, budgetApprovalId, executionPolicy, creativeAttempt, actor]
      properties:
        schemaVersion:
          const: v1
        shotSpecRevisionId:
          type: string
          format: uuid
        promptSnapshotId:
          type: string
          format: uuid
        generationProfileRevisionId:
          type: string
          format: uuid
        generationPlanId:
          type: string
          format: uuid
        routeSnapshot:
          $ref: "#/components/schemas/ModelRouteSnapshot"
        budgetApprovalId:
          type: string
          format: uuid
        executionPolicy:
          $ref: "#/components/schemas/ExecutionPolicy"
        creativeAttempt:
          type: integer
          minimum: 1
          maximum: 2
        fallbackReasonCode:
          type: string
        actor:
          $ref: "#/components/schemas/Actor"
...
    GenerationRun:
      type: object
      required:
        [runId, shotSpecRevisionId, runSpecDigest, creativeAttempt, state, traceId, createdAt]
      properties:
        runId:
          type: string
          format: uuid
        shotSpecRevisionId:
          type: string
          format: uuid
        runSpecDigest:
          $ref: "#/components/schemas/SHA256"
        creativeAttempt:
          type: integer
          minimum: 1
        state:
          type: string
          enum:
            [DRAFT, VALIDATED, QUEUED, RUNNING, PAUSED, UNKNOWN, RECONCILING, REQUIRES_ACTION, CANCEL_REQUESTED, CANCELLED, SUCCEEDED, FAILED]
        failureClass:
          type: string
          enum:
            [INFRASTRUCTURE, TRANSIENT, CREATIVE, INPUT, LICENSE, BUDGET, AUTHENTICATION, PERMISSION, QUOTA, RATE_LIMIT, CONTENT_SAFETY, REGION, MODEL]
        failureCode:
          type: string
        traceId:
          type: string
        createdAt:
          type: string
          format: date-time
...
    FreshnessImpact:
      type: object
      required: [affectedType, affectedRevisionId, state, reasonCode]
      properties:
        affectedType:
          type: string
        affectedRevisionId:
          type: string
          format: uuid
        state:
          type: string
          enum: [STALE, REVALIDATED, REGENERATED]
        reasonCode:
          type: string
...